backend/requirements.txt
Flask
google-generativeai
youtube-transcript-api
flask-cors
pytube

